requirements.txt
certifi==2019.3.9
dj-database-url==0.5.0
Django==2.1.7
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
whitenoise==4.1.2
wincertstore==0.2
